added optional hooks
